Allow hiding the language button
The language button at the top of the page is always displayed
even when the page is not available in other languages. This change
allows hiding the button. By default the language is button will
be showed, and it maybe disabled if the page is not availabe in
different languages/variants.

Bug: T130225
Change-Id: I38bf8695e5f764cb2fd66758694018edec854849