Update patch set 1
Patch Set 1: Code-Review+1

Wondering if we should default this to true. For a wiki that is being setup it would help them troubleshoot whether languages are broken or simply not available.

Given mediawiki.org and wikitech are the only ones I know that would need to use this I feel this the majority.

Patch-set: 1
Label: Code-Review=+1