Beta: Vertically align the page actions with the title
When the tagline is present on the page, it is displayed
below the title. Since the page actions were absolutely
positioned relative to the bottom of the container on
tablet-sized screens, that would mean that the page actoins
would be pulled down. Positioning relative to the top of the
container makes sure that the page actions are always in line
with the header.

Bug: T131050
Change-Id: I39668a9528a9f3906509014ca7d2c1ba0f2ebdbb