Update patch set 4
Patch Set 4:

OutputPage::removeModule() is not what you want. The module load is a side-effect of code somewhere deciding to build a TOC and add it to the page. One should not try to unload the module, but rather ensure the TOC isn't created in the first place. Modules may be renamed. And code does not expect its actions to be undone. It's fragile and wasteful all around.

Doing it in OutputPage also wouldn't cover cases that use the API to retrieve things for mobile. This information should be cached in ParserOutput indeed.

Note I haven't looked at the hook being used here in this commit. Just a reply to the previous comment.

Patch-set: 4
Label: Verified=0