Update patch set 8
Patch Set 8:

I'm unsure whether to +1 this. The change LGTM, though I'm still not sure about the rationale behind deleting the contents of a file but not the file itself – @Bmansurov provides some extra detail as part of T131757 – and I've given it a rough testing locally.

I'd be happier if:

* I understood the gotcha(s) that we're avoiding by not deleting mobile.special.mobilediff.scripts/mobilediff.js
* Jkatz and Dr0ptp4kt confirmed the data that's being logged by these schemas really isn't being used

Patch-set: 8
Label: Verified=0