Update patch set 2
Patch Set 2:

So I was wondering how we should document this change, as 'is_js_only' is kind of becoming a required field for this feature to work. Should we note somewhere saying that this key is required so that new page actions don't skip it? Any ideas?

Patch-set: 2
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0