Update patch set 10
Patch Set 10: Code-Review+2

On Firefox 8.0 / Android 3.2.1 Motorola tablet the space is reserved for ToC but it's not visible. The browser settings indicate that the JS is enabled, but it doesn't seem like so as sections cannot be collapsed or the main menu cannot be opened without leaving the page. This is an old device, so may not be a big deal as nothing is broken. The patch works fine on the stock browser on the same device though.

The patch looks good on other devices I used to test.

Patch-set: 10
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2