Correctly render the watchlist icon
When OOJS is loaded on a page, it overrides the .mw-ui-icon-watchlist
inside the main menu. Make the CSS rule to be more specific to prevent
this. This change can be reverted once T131931 is fixed. The problem
can be replicated by clicking on the echo (bell) icon and then
clicking on the main menu.

Bug: T133156
Change-Id: I3b44b6dbcd39bd7e9abec64c66184ededadb8da8