Update patch set 1
Patch Set 1:

Main menu is rendered via JS or on a special page so should not have caching issues.

Ideally this icon should be generated via the ResourceLoaderImage module (the fact it doesn't is a bug  https://phabricator.wikimedia.org/T95224 )

I would prefer we stay compatible with ResourceLoaderImage module to avoid complicating the future fix for that bug.

The existing fix also assumes we may never want to use the icon elsewhere outside of the menu.

Patch-set: 1