Update patch set 1
Patch Set 1:

This one is a little tricky because although the menu is being generated with JS, the data is coming from the backend: https://github.com/wikimedia/mediawiki-extensions-MobileFrontend/search?utf8=%E2%9C%93&q=wgMinervaMenuData that's why I thought the menu items would be cached. Do you think that's not the case?

> I would prefer we stay compatible with ResourceLoaderImage module to avoid complicating the future fix for that bug.
Are you suggesting that we fix the above bug you mentioned? Renaming the icon won't satisfy this requirement either.

> The existing fix also assumes we may never want to use the icon elsewhere outside of the menu.
This is true, and we have a pretty good chance of not using the watchlist icon outside menu until T131931 is fixed. We haven't used that icon in another place for so long, and I don't expect us to make drastic changes to the UI in the coming months.

Patch-set: 1
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0