Update patch set 1
Patch Set 1: Code-Review-1

This one is a little tricky because although the menu is being generated with JS, the data is coming from the backend: https://github.com/wikimedia/mediawiki-extensions-MobileFrontend/search?utf8=%E2%9C%93&q=wgMinervaMenuData that's why I thought the menu items would be cached. Do you think that's not the case?

We should check this on a cached HTML. Since the template is generated in JS via ResourceLoader (5min cache) I think we are okay.

> I would prefer we stay compatible with ResourceLoaderImage module to avoid complicating the future fix for that bug.
Are you suggesting that we fix the above bug you mentioned? Renaming the icon won't satisfy this requirement either.

In a follow up since we are here in this piece of code, this might be a good time.

> The existing fix also assumes we may never want to use the icon elsewhere outside of the menu.
>This is true, and we have a pretty good chance of not using the watchlist icon outside menu until T131931 is fixed. We haven't used that icon in another place for so long, and I don't expect us to make drastic changes to the UI in the coming months.

Let's follow your suggestion in the parent bug and simply prefix all these menu icons with mf- to keep us safe in future given the main menu is so prominent (we may later want to do this for page actions too but let's hold off on that for the time being)

Patch-set: 1