Update patch set 1
Patch Set 1:

If you can show me a browser that doesn't support nav that is in the wild (ie. you are able to download it) and show me a fix for that particular browser we have a discussion but right now this just seems hypothetical and YAGNI.

I'm also happy for us to explicitly say we will not support IE8 instead of merging this patch but right now it feels like we are making up an issue that isn't really there.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0