Add $wgMFStripResponsiveImages setting, defaulting on
Introduces $wgMFStripResponsiveImages setting, defaulting to true to
avoid changing current behavior. Setting to false will restore the
previous default behavior of not removing srcset attributes on the
mobile view.

Note that this does not add to the parser cache has key; a second
commit will add this as there is some controversity over it. As with
the current behavior, srcset may be stripped from the parser cache
on mobile *renderings* but still *served to desktop* or vice versa.

That will not affect production renderings on Wikimedia at the moment
due to a separate config hack which already splits the parser cache.

Bug: T134115
Depends-On: Iad06a37015d803c39f4a5ea69cbffc38a70cf6f3
Change-Id: Icc1246e2b109e48ec5d178c146a779980ce6c196