Update patch set 2
Patch Set 2:

> I'm not sure how I feel about this.
> If a page has multiple svgs then a slow connection loses the benefit of the srcset removal.

You still get the benefit of the srcset removal on all the JPEG photos. :)

We should probably do a quick survey of the file sizes for common maps & icons and whatnot to confirm or deny my hunch that they usually compress well enough at 1.5x and 2x.

> Using a whitelist seems like a hacky way around this (could we not just do something like Image->isSVG?

There is file->isVectorized(), or the actual file->getMimeType to be more specific about source file type... but you can't get at MediaTransformOutput->file property currently as it's protected and not exposed via the hook. Would need to patch core to either expose the file via an accessor or on the hook.

> I know a while back we considered actually embedding real SVGs in the page. I remember Ori was pushing for that but I never saw a write up explaining benefits/cons of doing so.
> Is this worth reigniting?

Yes, yes, and yes. Quick summary:
* upside: many SVGs are smaller than their PNG rasterizations
* upside: native-resolution rendering, even on 3x and 4x displays, without further increase in file size
* downside: some SVGs are actually HUGE, for instance super-detailed maps
* downside: some SVGs render very slowly even if not huge, for instance with heavy filter usage
* downside: browser compatibility issues
* downside: difficulty in availability of fonts for text & i18n

Patch-set: 2
Label: Verified=0