Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

If the toggling JavaScript fails to load or has a bug the content of sections is inaccessible.

I've heard rumours that the research Tilman is doing shows engagement is higher when these are open by default so let's wait for that to finish before meddling with this. I suspect we'll be able to leave them open by default and avoid this issue.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1