Update patch set 8
Patch Set 8: Code-Review-1

The current behavior is by design. The assumption is that it is not critical for a change to this configuration option to apply everywhere immediately at the cost of (effectively) wiping out the parser cache. Brion, do you think that is the wrong call to make?

Patch-set: 8
Reviewer: Ori.livneh <255@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1