Add categories on a new line
Currently, the categories are in one line and are appended directly
after the current content of the page (not on a new line). It would
be more readable for the users editing wikitext, if each category
is added on a new line (including the first added one).

This change changes the behaviour of the category overlay to do that.

Bug: T130503
Change-Id: I70fe2baedc8b03f175fa34961228000140c78653