Hygiene: Move page issues initialisation code to skins.minerva.scripts
The loading of mobile.issues should not have side effects and should
be reusable. There is no longer any reason for this code to be outside
skins.minerva.scripts

Additional changes:
* Remove outdated code which attempts but fails to add an issues banner
to the editor preview. This is no longer shown and hasn't been for some time
so one can only assume is not necessary at current time.

Change-Id: I069bc7d71e4f11928d544efbcbc3c5b4c12491d7