Update patch set 10
Patch Set 10: Code-Review+1

QAd, works fine. References with JS disabled send you to the Special page with the HTML, and clicking on a reference actually highlights the reference all with the hashtag and HTML only.

It's missing a browser test but I guess that can be done as a separate patch.

Patch-set: 10
Reviewer: Jhernandez <2027@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1