Update patch set 1
Patch Set 1: Code-Review-1

The impact of removing images is extremely jarring on a mobile phone and was specifically requested for this reason by our design Nirzar. Note that Chrome blocks rendering during a scroll event so this will result in images appearing instantly in a jerky fashion. This should be discussed more on the task.

Patch-set: 1
Label: Code-Review=-1