Disable MobileFrontendHooksTest::testOnBeforePageDisplay
The test tries to assert that no Vary headers are set in OutputPage,
which is not something a test should be doing without lots of mocking
(OutputPage is not isolated very well and MobileFrontend is certainly
not the only component messing with headers).

This is breaking all CI, can be reenabled after it is fixed.

Bug: T135866
Change-Id: I4631142df2cd737351bd03d56b238183787c63b0