Update patch set 5
Patch Set 5:

I'm a little confused. We have this problem already on mobile and its documented in this bug: https://phabricator.wikimedia.org/T126825 and it is blocked on Tilman's research.

collapsing with CSS would break references altogether for users who fail to download JavaScript due to faulty connection.

Patch-set: 5