Update patch set 5
Patch Set 5:

The link you have given goes against this patch actually. If we're not sure if collapsing is good we should not do this patch.

As for when JS fails to load, this is true for anything else on the page, isn't it? For example, the watchstar code may fail to load thus making watchstar unusable.

Patch-set: 5