Update patch set 5
Patch Set 5:

>The link you have given goes against this patch actually. If we're not sure if collapsing is good we should not do this patch.
We don't know the answer to this yet.

> As for when JS fails to load, this is true for anything else on the page, isn't it? For example, the watchstar code may fail to load thus making watchstar unusable.
Yes, but reading content is important than using a watchstar. Hiding a section that cannot be opened that contains content is a bad thing to do. Can we take this bikeshed into a bug or follow up patch? This is not a blocker imo given mobile already exercises this issue in far more terrible ways.

Patch-set: 5
Label: Verified=0