Update patch set 2
Patch Set 2:

According to mw.viewport documentation, the threshold argument is based on proximity to the edge of the viewport. Not the top of the screen.

As such, passing 1.5x the height of the window, means an image will load even if it is not on the current screen, and also not in the next "full-page scroll", but in the second full-page scroll away from the current scroll position. That seems relatively far?

I think the idea was to extend the lookahead to the next half screen full. And presumably you thought offset is measured from the top of the screen?

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0