Update patch set 7
Patch Set 7:

An attribute lookup is far cheaper than a DOM search for a class element.  We can of course cache the result but then one would wonder why we don't use a DOM attribute in the first place.

What is your concern about adding the data attribute flag? Since the pattern is repeated it gzips great.

Patch-set: 7
Label: Verified=0