Update patch set 8
Patch Set 8:

I'm mainly concerned that adding the data attribute puts extra load on the server, while we could ship this exercise to the client. It's also only used when JS is available. We're penalising non-JS users with the change.

Also, we should only change the DOM if it is absolutely necessary, such as when we want something to work for all users (JS/non-JS). In this case the data attribute is signalling the front-end that the section is OK to be collapsed, which can be done without adding the attribute in the back-end.

Patch-set: 8