Update patch set 8
Patch Set 8:

> I'm mainly concerned that adding the data attribute puts extra load on the server, while we could ship this exercise to the client. It's also only used when JS is available. We're penalising non-JS users with the change.

How so? After gzipping the HTML change is minor if non-existent. In MobileFormatter we are already looking up the section so adding a data attribute to it is cheap.

> Also, we should only change the DOM if it is absolutely necessary, such as when we want something to work for all users (JS/non-JS). In this case the data attribute is signalling the front-end that the section is OK to be collapsed, which can be done without adding the attribute in the back-end.

It can but shouldn't as I said in my last comment. There's no need to compute it and query for elements because of the above. It's unnecessary and far more error prone as we'd have to manage it in 2 code languages. What if the references class changes?  We would have to change the logic in MobileFormatter and JS.

Patch-set: 8