Update patch set 8
Patch Set 8:

> Also, we should only change the DOM if it is absolutely necessary, such as when we want something to work for all users (JS/non-JS).

I disagree with this statement.

We're querying and manipulating the DOM in MobileFormatter so that the majority of users, i.e.
users with Resource Loader capable UAs, don't have to, while adding a minimal amount of metadata
to users that can't use it.

Note well that the majority of responses are being served from Varnish rather than from the
application servers.

So, while we are somewhat impacting the performance of our server-side application, we're doing it
so that we don't impact the performance of our client-side application //nearly as much//.

This is a trade-off.

You must also consider that we can control the performance characteristics of our server-side
application far more easily than we can our client-side application – consider the variety of user
agents [0], for example, all with varying resource available to them.

> After gzipping the HTML change is minor if non-existent.

We'll ignore the "if non-existent" bit… ;)

Note that gzip probably won't have a large impact here as the pattern will be applied to one
element for the majority of pages that use the <references /> tag.

[0]: http://datavis.wmflabs.org/agents/

Patch-set: 8
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0