Update patch set 8
Patch Set 8:

I'm not convinced. The only thing that the front-end needs to do is to query for references and get it's parent div and collapse it. We have millions (not really) queries in JS, this new one will not be a performance degrader. That said, if you guys are happy with the change, I'm happy.

Patch-set: 8