Update patch set 8
Patch Set 8:

> I was mainly talking about the HTML transformation that we need to do and not about the HTML size. Besides, as I said earlier, what will happen to cached pages?

For those pages that haven't been transformed (read: haven't fallen out of the cache yet)? Nothing! The attribute won't exist and the new behaviour won't kick in. This feature will roll out over a period of seven days – the amount of time it takes cache hits to drop to ~0%.

Patch-set: 8