Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Also, what's the motivation behind this change? Are we going to have different values depending on the wiki? Are we going to change this value often? Is it needed for an A/B test? I'd appreciate it if you could briefly talk about it in the commit message.

Patch-set: 2