Do not remove bottom border of fields in signup form
On a narrow screen, the signup form gets a "stacked" look with the
fields sharing horizontal borders. This used to include removing the
bottom border, which doesn't work so well with the new HTMLForm-based
signup where error messages can appear under the fields. The patch
restores the border; the fields have margin-top: -1px anyway so
the borders will still overlap.

Bug: T135775
Change-Id: I8a4732cf54c47112baade02126d8c81738bb1403