Don't mark document as reference section
When formatting the main page or any special page, the entire document
will be transformed in one pass, rather than section by section -- it's
likely there won't be heading tags to section the document by.

If there exists a list of references on the page, and the On Demand
References feature is enabled, then the list of references will be
transformed and its parent section marked as a "reference container."
Per the above, this transformation will be done on the document (i.e.
the DOMDocument instance), and when #setAttribute is invoked on document
(i.e. DOMDocument#setAttribute) a fatal error will be triggered.

This change is the smallest possible change to stop this fatal error
from being triggered.

Bug: T137375
Change-Id: I086f7fa45fb96b1abcba5c78d8eb25fdd36b19f7