Update patch set 1
Patch Set 1:

The nasty thing about the config hack (which you pointed out and corrected -- and thank you for that) was the $wgRenderHashAppend bit, because without it the code in MFE would make a mess of the parser cache. But varying the option value based on whether or not the request was made to the API doesn't seem so gross, because it's a pretty Wikimedia-specific requirement. So how about simply moving the configuration back into the EnterMobileMode hook handler, and making it "$wgMFStripResponsiveImages = defined( 'MW_API' );"? It's not pretty, but it avoids the cost of complicating the code for one special case. What do you think?

Patch-set: 1