Update patch set 4
Patch Set 4: Code-Review-1

(5 comments)

Making the Toggle class load references on demand and add them to the section breaks SRP <https://en.wikipedia.org/wiki/Single_responsibility_principle>. There are many examples of this in our codebase but, fortunately, the Toggle class already has a mechanism for alerting other systems that the user has toggled a section: the section-toggled event.

There should exist an event listener that loads a section's references on demand and adds them to the section and it should be distinct from the Toggle class. That way, if we ever need to disable/remove on-demand reference loading, then the Toggle class doesn't need to change.

Patch-set: 4
Label: Code-Review=-1