Update patch set 4
Patch Set 4:

OK, that makes sense. How would I go about not toggling a section until after references are loaded? As things currently stand, toggling happens before the aforementioned event is emitted. The loading can happen in another file/module but there would be a FOUC as the section contents would need to be replaced by a spinner.

One solution I can think of is to emit another event once the section content is ready to be displayed, then the Toggle which will be listening out for the event class would just display the content.

@phuedx, what do you think?

Patch-set: 4