Update patch set 4
Patch Set 4:

Again. I think that since the only listener of the event isn't at all concerned with these details, then we have an opportunity to tweak the firing of the event to what we need – though we should make sure we document it well in the source :)

If we fire the event //before// the section is toggled, then the handler can replace the contents of the section – with a spinner in our concrete – before the section is displayed.

What do you think?

Patch-set: 4