Update patch set 4
Patch Set 4:

A problem with that approach is that we want to keep the section content for a later use in case the API call fails. So, we'll have to cache it and insert back when needed. The downside is that we'll be doing unnecessary DOM manipulations. My approach was that we don't toggle the section at all and just show the spinner until we get a response from the API, then replace the content of the section.

Patch-set: 4
Label: Verified=0