Fix reflow from lazy loaded images improvements
See T135430#2384461

The link that wraps all images is an inline element and if the children is not
a block it will change it's size. Before this didn't happen because we had
overflow:hidden which hid the extra space keeping the element unchanged.

This patch makes the placeholder a block, like the inside image that will load
is, making it unnecessary to bring back overflow:hidden

Bug: T135430
Change-Id: I0bd6676a523669d47fede7714db3ab4b7b24fa4a