Update patch set 22
Patch Set 22: Code-Review-1

(1 comment)

Browser test failures are valid.
I'm seeing

TypeError: this.referencesSetup is not a function TypeError: this.referencesSetup is not a function(…)

resources/skins.minerva.scripts/preInit.js

This will not be true when shouldLazyLoadReferences is false.

I would recommend moving this code out of preInit for our sanity.

Patch-set: 22
Label: Code-Review=-1