Un-break page actions on User pages
Remove the user page specific rule so that page actions styling
can style the heading holder. The change makes sure that there is
enough padding bottom of heading holder for page actions to occupy.
On small screens page actions container is absolutely positioned,
thus it requires space to be reserved.

A downside of this approach is that the space is always reserved even
when the page actions are invisible. This may happen when all page
actions are JS-only and the browser doesn't support JS.

Alternatively, we could introduce a new class for heading holder
which would let us know whether all page actions are JS only or not.
This will allow us to not reserve space for page actions when they
are invisible.

Bug: T137054
Change-Id: Id0dfec608856eca936d4ddbccc56d88bc53888c9