Hygiene: Separate tagline from heading
The tagline should be part of the postheadinghtml - as it is not a heading
but appears after the heading.

This caused me lots of confusion when trying to understand where and how
the tagline was generated as a tagline is not the same as a heading and is more
closely associated with the user page actions  - so I'm keen to
correct this mental model.

Additional changes:
* Remove redundant SkinMinervaBeta::prepareHeaderAndFooter -
it duplicates code in SkinMinerva.

Change-Id: I4763007d91ae49e7b5098c2ac5d5e3a6bab80011