Update patch set 1
Patch Set 1: Code-Review-1

We should take more time over this.

If we want to do a quick fix for production with Nirzar's consent we should simply make the page actions display:relative.

This is in beta thus we should take time to get this correct and to be correct we should move the page actions below the heading. It should be positioned at the end and we should revisit how we float it to the right in tablet mode.

The use of padding will just delay this problem to a future date when we want to make further changes to the header.

Patch-set: 1