Load all reference lists inside a reference section
Take into account that a references section may have
subsections and that each of these has to be loaded.

Changes:
* When opening a reference section replace any reference
list placeholders with their HTML.
* lazyLoadReferences now returns a $.Deferred object to allow testing
* New Skin test checks that a lazy loaded reference section behaves correctly
* Use template for readability to ship a suitable complex example which covers all the test
scenarios (two references lists in a single section, references in subsection and references
in lead of a section.
* getReferencesSection and getReferencesSections renamed to getReferencesList and
 getReferencesLists to capture better what is needed. We are only interested in loading the
lists of references not the entire html of the section

Bug: T138792
Change-Id: I903ec8e109f0de35ca001b8511a1b990fb4bff02