Update patch set 1
Patch Set 1: Code-Review-1

Can we implement a more general solution? Add a negative margin left to the last page action li and a negative margin right to the first page action li? Right now the language button is correctly aligned with the left border, but the edit icon is not with the right border.

Patch-set: 1
Label: Code-Review=-1