Update patch set 1
Patch Set 1:

The last patch aligns the edit icon to the right.

In terms of doing this in a generic fashion, given that we already have set a precedent with the use of the #language-switcher selector, I have posted a follow up that leans on first-child and last-child selectors to achieve this.

Patch-set: 1
Label: Verified=0