Update patch set 18
Patch Set 18:

I think that you're right about splitting up the patch into parts to make it a little easier to consume.

As suggested, I've set MFDisplayWikibaseDescriptions to true by default. 

Regarding your last comment, you were at the meeting where we discussed that while it is true that UI consistency is a target so is not overloading the pageterms API. Having a flag per API query may seem like overkill but so does turning everything off in the event that we add a new feature and it's the straw that breaks the camel's back.

Patch-set: 18
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0