Deprecate $wgMFDisplayWikibaseDescription
... and $wgMFDisplayWikibaseDescriptionAsTaglines.

Add $wgMFDisplayWikibaseDescriptions, a map of internal feature name to
the flag that determines whether or not the Wikibase description should
be shown. Currently, the internal feature names are as follows:

* search, and
* tagline

The former is the more general, flagging the display of Wikibase
descriptions in search results, including nearby search, and watchlists.

Bug: T138788
Change-Id: Iecc2224ca000530c286cd82a7b7db0383746cdee