Update patch set 1
Patch Set 1:

The idea was to improve the maintenance of the code that is here by consolidating it with the bulk of the mobile-oriented frontend code.  But you make a good point that this would result in shipping code that desktop clients have no use for.  It doesn't seem like the MobileApp extension is doing much, but maybe there is a case for its continued existence after all.  

One possible alternative I've had on the back burner is to write a bot that watches for changes to the MF less files we pull in, and immediately runs the update script and create a patch for the app when a change is merged.  That would at least save us from running the script manually as often as we happen to think of it.  That said, a good old cron job would probably be more reliable and get the job done almost as well.

Patch-set: 1
Reviewer: Mholloway <2478@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0