Update patch set 4
Patch Set 4: Code-Review+1

As engineers we have a responsibility to clean up where ever we get an opportunity and make our code better. It's not going to fix itself. Adding an identical PNG in two places (does that not feel wrong to you?) just adds to the problem.

Given this was trivial I've fixed it myself rather than arguing the point further.

Patch-set: 4