Do not apply horizontal padding on basic search interface
In I7b0dc769282c8550998dd29448a07105a73a1507 the search
form we show Opera Mini users was given a left padding of 2em
to make space for the magnifying glass. This is no longer needed
as it is inherited from the generic rule. Right padding is however retained
at 0.1em as previously to maximise space for typing and the placeholder (which
is taken away by the search icon)

Up to 240px resolution we will not show the search icon in the search
input at all given it takes up precious real estate.

Bug: T140490
Change-Id: I367833c2fa648b5673ab9dcabb5bc4dbf5a01683