Reset distance between search icon and placeholder
... on the language switcher overlay search input

In WebKit, there are additional pseudo-elements added to a search input
that are still rendered when the element's appearance is "none". These
elements increase the distance between the custom search icon and the
placeholder text. In Gecko, however, no such pseudo-elements are added.

Correctly hide all of the pseudo-elements added to the language switcher
overlay's search input so that all rendering engines render the element
equally.

Bug: T139928
Change-Id: I3cec13082380a2c866122680509fd8514e40ca20