Update patch set 1
Patch Set 1:

I'm not sure I see the value in adding a config variable/feature flag for a button we're phasing out. In //principle//, I agree with your approach here, but in this specific instance, I think it might be unnecessary and could lead to confusion. Specifically, somebody checking git log and finding a new config var for a button that's disabled in beta could be very confusing.

Patch-set: 1