Update patch set 1
Patch Set 1:

> Cannot help but wonder if this should be in the global reset. Another option is to remove the search type from the input.

While that would certainly solve the problem, I think that it's the wrong solution. In this context, <input type="search" ... /> is more semantic than the alternative. We simply need to reset WebKit slightly more aggressively than we once thought.

> Any reason we wouldnt want to do that (note this css rule applies to all inputs not input[type-search]?

This change fixes the problem and hopefully explains why the solution is ideal. The follow-on changes either tidy up, either by making things a little clearer or relocating code that can be shared.

With that in mind, I definitely want this in a reset, but I'll move it in a follow-on change.

Patch-set: 1
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0