Update patch set 6
Patch Set 6: Code-Review-1

Clicking on the old button (which is at the bottom of the page) shows the tooltip correctly. However, clicking on the same button again doesn't do anything? I think we should display the tooltip as many times as the user clicks on the bottom button. The user may miss the first tooltip.

Patch-set: 6
Label: Code-Review=-1