Update patch set 1
Patch Set 1: Code-Review-1

Given that it is clear this solution is very brittle - Im not convinced this is the right solution. Weve only repurposed config variables so I suspect if something is broken there is an issue at the config level not the code level. If this really is needed please add a qunit test to avoid this breaking yet again and explain better what it is doing.

Patch-set: 1
Label: Code-Review=-1