Update patch set 5
Patch Set 5:

I put WIP on there because it's still missing the PHPUnit test and I wanted to be clear that it shouldn't be merged (hence the "TODO" in the commit message as well). It'll be removed with the next patch set.

Patch-set: 5
Reviewer: Jhobs <1948@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0