Update patch set 8
Patch Set 8: Code-Review-1

IF we are worried about the delay we should just do the simplest thing and load mobile.contentOverlays at the start of page load.

IMO we are really getting stuck in the weeds here given this is only in beta and we've had no designer input on what should happen when the button is tapped a second time (maybe the button should disappear!?!) . I would suggest we all talk after standup with Nirzar and get this wrapped up rather than do this asynchronously and slowly.

Patch-set: 8
Label: Code-Review=-1