Update patch set 10
Patch Set 10: Code-Review+1

Would be useful to update the README for this config variable to show what we are changing. The only visible difference I see is the edit and watch icon are reserved in tablet and page actions appear horizontal in mobile mode. Would be good to make this clear in the README in case there are other changes.

Patch-set: 10
Label: Code-Review=+1