Use InfiniteScroll for the CategoryOverlay
Currently, the list of ctagories is fixed to 50. Even if 50 sounds
like a huge amount of categories, it is possible to have more than
that. Until now, the user was not able to see the other categories
after the first 50.

This change implements the use of InfiniteScroll to load the next
50 catgeories, after the user eached the end of the list of categories.

Change-Id: Ibe1e0db0307e7a9d89c356777b56fc0c7263360f