Combine scroll handlers into the global one
In order to keep the number of handlers attached on scroll to the window, let's
use the events in mw.mobileFrontend so that there is only one handler attached
to the window.

This moves the handlers on back to top and the infinite scrolling component to
use the global event handler (using 'scroll' when debounced or
'scroll:throttled' when needed)

Also fixes the infinite scrolling unit tests to rely on the global event
emitter instead of in syncronous window bound scroll events.

Bug: T124870
Change-Id: Ic21dc04389cab065c6dd52968689f72c045939ab