Hide empty li elements according to tidy rules (like in mediawiki/core)
It's not the best idea to copy the css rule from core, but the better idea
as including mediawiki.skinning.content into mobile web. Unfortunately, this
increases maintenance costs and is a hack only (that's why it lives in "hacks")
and should be removed, once we found a better handling for the same core/mobile css
rules, without including huge modules with mostly unused rules.

This change synchronizes the handling of empty li elements ("* \n" in wikitext), which
were hidden in desktop and visible in mobile. MobileFrontend now hides these
elements, too.

Bug: T129375
Change-Id: I05f9241a6f930f50271653f0b01ee030f760d00a