Fix the primary navigation background
It should not be white. The regression was introduced in
I41bf905d248fb389edbe100a78309e0f523cae58.

Loosen the selector so that the CSS rule for background color
is overridden when animations kick in.o

In beta: make sure the page center background color is the same
as the page background color so that the primary navigation
doesn't overlap the page center when closing the navigation drawer.

Bug: T144035
Change-Id: Icf330d74bb6647dd1006e687e116b8d80a99c9a1