Update patch set 5
Patch Set 5: Code-Review+1

(1 comment)

This LGTM.

I was able to reproduce the error by adding the following to LocalSettings.php.

    $wgDefaultSkin = 'minerva';
    $wgMFEnableMinervaBetaFeature = false;

I briefly removed the above and enabled the beta feature for a test user on my local wiki and ran the following tests:

1.
    $wgDefaultSkin = 'minerva';
    $wgMFEnableMinervaBetaFeature = false;

Special:Preferences should load and report Minerva as the chosen skin. Minerva and Vector should appear in the list of skins. ✔

2.
    $wgDefaultSkin = 'vector';
    $wgMFEnableMinervaBetaFeature = false;

Special:Preferences should load and report Vector as the chosen skin. Minerva shouldn't appear in the list of skins. ✔

3.
    $wgDefaultSkin = 'vector';
    $wgMFEnableMinervaBetaFeature = true;

Special:Preferences should load and report Vector as the chosen skin. Minerva and Vector should appear in the list of skins. ✔

4.
    $wgDefaultSkin = 'minerva';
    $wgMFEnableMinervaBetaFeature = true;

Special:Preferences should load and report Minerva as the chosen skin. Minerva and Vector should appear in the list of skins. ✔

5.
    //$wgDefaultSkin = 'minerva'; **NOT SET DELIBERATELY**
    $wgMFEnableMinervaBetaFeature = true;

Special:Preferences should load and report Vector as the chosen skin. Minerva and Vector should appear in the list of skins. ✔

Patch-set: 5