Treat list items as inline
By default an `li` element is inline. So any img tags inside
that are also inline should also be inline.

When a `li` tag is not inline it means classes have been applied
to it and thus any `img` tags inside the `li` element should also
be explicitly styled via classes.

Since we copy classes across to the placeholder, it should be safe
to assume any img inside a list item should be treated as inline.

Bug: T144567
Change-Id: I911f305fa130e60991e830708d6680599870cc13