Update patch set 1
Patch Set 1: Code-Review-1

-1 for your attention.

There's an opportunity to make the client as simple as possible by shifting more work to the server. We shouldn't have to traverse the DOM in a scroll handler.

We currently extract the width and height attributes from the <img> in order to style the placeholder [0]. We should also try to extract them from the style attribute, if one is set.

https://github.com/wikimedia/mediawiki-extensions-MobileFrontend/blob/master/includes/MobileFormatter.php#L257-L258

Patch-set: 1
Reviewer: Phuedx <1196@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1